Fix import wallet (#592)
* added bridge methods for wallet get,add,delete,update apis

* test

* update config

* ios bridge methods

* feature: restore redux store from json dump

* refactor: restore data store into it's own function

* tests: restore tests

* glow fix

* fix: flow issue with an any type

* refactor: use restoreDataStore generator

* style: remove comment & unused variable

* refactor: restore store methods to account for wallet structure

* feature: hydrate saga after successful restore

* fix: issue with async safe storage methods

* refactor: restore flow

* update stuff

* feature: new apis for import export

* adds todo comments for restore backup flow

* updates

* 🤷🏼🤜🏼

* fix: flow restore is functional

* feature: adds backup wallet changes

* fix: makes wallet file name a constant for reuse

* feature: unzipWalletRestoreFile scaffold for android

* refactor: move unzip from native to js layer

* call export API

* comment ensureVcxInitSuccess

* feature: adds salt.json file to backup flow

* comment ensureVcxInit

* adding comments for backup flow

* refactor(backup): zip if now done in js layer

* added missing string

* Changed the name of the parameters to decryptionKey for the decrypt wallet method

* Fixed the path to the file for ios to be the path that exists

* Updating the snapshots

* fix import API integration

* refactor: adds additional parameters to importWallet methods

- adds shutdownVcx method stub to JS bridge

* feature: simple vcx init stub

* fix isLockEnable and isTouchId

* fix: syntax

* fix: wallet key

* fix: comment out importWallet method for now

* fix: remove ensureVcxInitSuccess from restore

* feature: adds new bridge methods to restore-store

* fix: change walletKey back to regular string for now

* fix: remove select for getUserOneTimeInfo in restore-store

* updating to new pod and aar

* update yarn lock

* fix issues

* updated pushnotification pushToken place of calling (#577)

* test

* update push notification token call

* code cleanup

* fix update push token

* updated with static data for restore flow

* removed consoles

* update rncxs.js update

* fixing error after merge

* update of bridge method for import wallet

* fix: simpleInit

* fix RNIndyModule

* fix: simpleInit method with proper parameters

* fix: updates vcxShutdown method

* value check in splash-screen for isLockEnabled

* WIP hydration

* Fixed the iOS bridge code for the importWallet api

* Now the shutdown api on iOS is working.

* Changed the return code to be a Number for the shutdown api

* Chaned the import statements to be correctly importing the functions

* Added the ensureVcxInitSuccess calls so that the init is successful before we try to access the wallet

* Added the ensureVcxInitSuccess calls to have the accessing of the wallet storage wait until the vcx init has occurred

* Now I can export a backup as well as import a backup on iOS

* fix importWallet

* Now the backup wallet flow works again and better logging statements in RNIndy.m and a clearcache target in the Makefile

* Now the export wallet and import wallet work all the way to restoring the pin after the wallet import

* remove ensureVcxInit from get/setWallet

* Now the wallet keys and values get saved to the wallet when starting out fresh and then the keys and values get imported correctly

* The checkPin now waits for the VCX_INIT_SUCCESS message

* Fixed the blocking on vcx_init_success

* fix push notification

* fix-connections

* remove unwanted commented code

* add history

* fix-kill-app issue

* fix importWallet

* fix lint errors and remove unwanted commented code

* Removed the temporary wallet storage object so that all we see are the errors in the logs for when the wallet has not yet been INITED

* Added a comment to indicate the problem with calling ensureVcxInitSuccess in the setPin function

* fix userOneTimeInfo issue

* hydration logic cleanup for alreadyInstalledNotFound

* code cleanup

* fix lock store circular dependency

* fix splash screen

* WIP: Made a single hydrate saga which handles whole normal app flow

* Changed the log message for when an error is returned from the libvcx call

* Remove loadHistory watcher

* Flow fixes for initialized flag

* Fix lint, tests, Normal flow with wallet CRUD APIs

* Fix connection store, Add bit more logging, Fix getWallet data fetching

* Fix issue with mapDispatchToProps on <Request /> component

* Fix secureGet to not throw, instead return null

* Fix import flow, Fix switch environment data store to wallet

* fix restoreDirectory path and push notification

* Fix salt generation for Android

* fix cross platform issue

* Removes commented out java code

* Removes extra dependency

* Address some comments made in push-notification.js

* Fix passphrase verification view

* add comments

* Fix push notification promise chain block

* fix multiple navigation issue

* Fix some broken tests

* fix lock setting kill app issue

* Fix connection-history test

* Fix set & get wallet bridge methods

+++ app/restore/restore-store.js
@@ -1,8 +1,7 @@
 // @flow
-import {
-  saveFileDocumentsDirectory,
-  decryptWalletFile,
-} from '../bridge/react-native-cxs/RNCxs'
+import { decryptWalletFile, copyToPath } from '../bridge/react-native-cxs/RNCxs'
+import { unzip } from 'react-native-zip-archive'
+import FCM from 'react-native-fcm'
 import { takeLatest, all, put, call, take, select } from 'redux-saga/effects'
 import type { CustomError } from '../common/type-common'
 import {
@@ -14,7 +13,10 @@ import {
   initialState,
   FILE_SAVE_ERROR_MESSAGE,
   DECRYPT_FAILED_MESSAGE,
+  RESTORE_RESET,
+  RESTORE_DATA_FAILED_MESSAGE,
 } from './type-restore'
+import { WALLET_FILE_NAME } from '../backup/type-backup'
 import type {
   SaveToAppDirectory,
   RestoreActions,
@@ -23,9 +25,15 @@ import type {
   RestoreStore,
 } from './type-restore'
 import RNFetchBlob from 'react-native-fetch-blob'
-import { getRestoreStatus } from '../store/store-selector'
-
+import { getRestoreStatus, getRestoreFileName } from '../store/store-selector'
+import { pinHash as generateKey } from '../lock/pin-hash'
+import { safeToDownloadSmsInvitation } from '../sms-pending-invitation/sms-pending-invitation-store'
 import { Platform } from 'react-native'
+import type { Store } from '../store/type-store.js'
+import { hydrate } from '../store/hydration-store'
+import { pushNotificationPermissionAction } from '../push-notification/push-notification-store'
+import { safeGet, safeSet } from '../services/storage'
+import { PIN_ENABLED_KEY, IN_RECOVERY } from '../lock/type-lock'
 
 export const saveFileToAppDirectory = (data: SaveToAppDirectory) => ({
   type: SAVE_FILE_TO_APP_DIRECTORY,
@@ -47,6 +55,10 @@ export const submitPassphrase = (passphrase: string) => ({
   passphrase,
 })
 
+export const resetRestore = () => ({
+  type: RESTORE_RESET,
+})
+
 export function* restoreFileDecrypt(
   action: RestoreSubmitPassphrase
 ): Generator<*, *, *> {
@@ -63,8 +75,61 @@ export function* restoreFileDecrypt(
     }
 
     yield put(restoreStatus(RestoreStatus.DECRYPTION_START))
-    yield call(decryptWalletFile, passphrase)
+    const { fs } = RNFetchBlob
+    const restoreZipFilePath = `${fs.dirs.DocumentDir}/restore.zip`
+    const restoreDirectoryPath = `${fs.dirs.DocumentDir}/restoreDirectory`
+    let walletFilePath = `${restoreDirectoryPath}/${WALLET_FILE_NAME}.wallet`
+    let restoreSaltPath = `${restoreDirectoryPath}/salt.json`
+    yield call(unzip, restoreZipFilePath, restoreDirectoryPath)
+    let restoreSalt = ''
+    try {
+      restoreSalt = yield call(fs.readFile, restoreSaltPath, 'utf8')
+    } catch (e) {
+      //TODO fix hack - for IOS the restoreDirectoryPath is having an extra level
+      // when backup is exported from android and imported in IOS
+      let fileName = yield select(getRestoreFileName)
+      let restorePath = `${restoreDirectoryPath}/${fileName.split('.zip')[0]}`
+
+      walletFilePath = `${restorePath}/${WALLET_FILE_NAME}.wallet`
+      restoreSaltPath = `${restorePath}/salt.json`
+
+      restoreSalt = yield call(fs.readFile, restoreSaltPath, 'utf8')
+    }
+
+    const parsedRestoreSalt = JSON.parse(restoreSalt)
+    const hashedPassphrase = yield call(
+      generateKey,
+      passphrase,
+      parsedRestoreSalt.salt
+    )
+    //TODO fix hack - for IOS need to do hash is having extra characters
+    // when doing cross platform export/import then it becomes incompatible
+    yield call(
+      decryptWalletFile,
+      walletFilePath,
+      hashedPassphrase.substring(0, 16)
+    )
     yield put(restoreStatus(RestoreStatus.FILE_DECRYPT_SUCCESS))
+
+    // since we have decrypted file successfully, now we restore data from wallet
+    // Need to set this here manually, because in normal flow this flag
+    // will be set when user sets pin code. However, since user is importing
+    // wallet file, we know that user has already enabled pin code
+    // so we set this flag manually
+    yield call(safeSet, PIN_ENABLED_KEY, 'true')
+    yield call(safeSet, IN_RECOVERY, 'true')
+    yield* hydrate()
+
+    //Push Notification permissions are asked when we do our first connection
+
+    //but in this case if connections are imported from backup then that case is missed
+    //since connection is already there
+    // so after push token update
+    //we need to do requestPermission or else push notifications won't come
+
+    yield call(FCM.requestPermissions)
+    yield put(pushNotificationPermissionAction(true))
+    yield put(restoreStatus(RestoreStatus.RESTORE_DATA_STORE_SUCCESS))
   } catch (e) {
     yield put(errorRestore(DECRYPT_FAILED_MESSAGE(e.message)))
   }
@@ -84,24 +149,25 @@ export function* saveZipFile(
     //For android device the content uri is of form "Content://com.android.providers.downloads.documents/document/223" which can be used directly
     //For IOS device uri is like "file:///private/var/mobile/Containers/Data/Application/A80FE508-BCED-4950-B9D0-8F1AA1E967B6/tmp/com.evernym.connectme.callcenter-Inbox/backup.zip"
     // which we need to split from /private to get the real path
+
+    //TODO move the copy logic for android to RNLayer
     if (Platform.OS === 'android') {
-      tempUri = decodeURIComponent(uri).split('/raw:')[1]
-        ? decodeURIComponent(uri).split('/raw:')[1]
-        : uri
-    } else {
-      tempUri = uri.split('/private')[1]
-    }
+      copyToPath(tempUri, destPath)
+      yield put(restoreStatus(RestoreStatus.fileSaved))
+    } else if (Platform.OS === 'ios') {
+      tempUri = uri.substr('file://'.length)
 
-    let sourceFileExists =
-      tempUri != null ? yield call(fs.exists, tempUri) : false
+      let sourceFileExists =
+        tempUri != null ? yield call(fs.exists, tempUri) : false
 
-    if (sourceFileExists) {
-      const destFileExists = yield call(fs.exists, destPath)
-      if (destFileExists) {
-        yield call(fs.unlink, destPath)
+      if (sourceFileExists) {
+        const destFileExists = yield call(fs.exists, destPath)
+        if (destFileExists) {
+          yield call(fs.unlink, destPath)
+        }
+
+        yield call(fs.cp, tempUri, destPath)
       }
-
-      yield call(fs.cp, tempUri, destPath)
       yield put(restoreStatus(RestoreStatus.fileSaved))
     } else {
       yield put(restoreStatus(RestoreStatus.FILE_SAVE_ERROR))
